+++ src/test/resources-noncompilable/com/puppycrawl/tools/checkstyle/checks/design/finalclass/InputFinalClassClassWithPrivateCtorWithNestedExtendingClass.java
@@ -1,6 +1,6 @@
-package com.puppycrawl.tools.checkstyle.checks.design;
+package com.puppycrawl.tools.checkstyle.checks.design.finalclass;
 
-public class InputClassWithPrivateCtorWithNestedExtendingClass {
+public class InputFinalClassClassWithPrivateCtorWithNestedExtendingClass {
     class A {
         private A() {}
         private class ExtendA extends A {}
@@ -9,7 +9,7 @@ public class InputClassWithPrivateCtorWithNestedExtendingClass {
     class B {
         private B() {}
         private class ExtendB extends
-                com.puppycrawl.tools.checkstyle.checks.design.InputClassWithPrivateCtorWithNestedExtendingClass.B {}
+                com.puppycrawl.tools.checkstyle.checks.design.finalclass.InputFinalClassClassWithPrivateCtorWithNestedExtendingClass.B {}
     }
 
     class C {
