+++ src/test/resources-noncompilable/com/puppycrawl/tools/checkstyle/checks/design/finalclass/InputFinalClassClassWithPrivateCtorWithNestedExtendingClassWithoutPackage.java
@@ -1,4 +1,4 @@
-public class InputClassWithPrivateCtorWithNestedExtendingClassWithoutPackage {
+public class InputFinalClassClassWithPrivateCtorWithNestedExtendingClassWithoutPackage {
     class A {
         private A() {}
         private class ExtendA extends A {}
